(deliver): Change mcost from float to double, because it's used as double but never as float.  This avoids converting the value of sector_mcost() to float and back.

+++ src/lib/update/deliver.c
@@ -55,7 +55,7 @@ deliver(struct sctstr *from, struct ichrstr *ip, int dir,
     int amt_moved;
     int amt_dst;
     int mobility;
-    float mcost;
+    double mcost;
     struct dchrstr *dp;
     int n;
